Build class lists in templates instead of presenters
Defining classes in templates makes more sense as then they're much
nearer to where they're used. Having to flit between two files to find
the css classes that are used in a given case is annoying.
