+++ src/Ds2013/Molecule/DateList/date_list_item.html.twig
@@ -1,14 +1,28 @@
 {% spaceless %}
     {% import _self as self %}
 
-    <li class="{{ date_list_item.cssClassesForLi }}">
-        {% if date_list_item.shouldShowLink %}
+    {% set li_classes = build_css_classes({
+        'date-list__page': true,
+        ('date-list__page--offset' ~ date_list_item.offset|abs): true,
+        'date-list__page--first': date_list_item.offset == -7,
+        'date-list__page--last': date_list_item.offset == 7,
+        'date-list__page--current': (date_list_item.offset == 0),
+    }) %}
+
+    {% set item_classes = build_css_classes({
+        'br-page-bg-onbg--hover br-page-linkhover-ontext--hover': date_list_item.isLink,
+        'br-box-page': date_list_item.offset == 0,
+        'text--faded': date_list_item.offset != 0 and not date_list_item.isLink,
+    }) %}
+
+    <li class="{{ li_classes }}">
+        {% if date_list_item.isLink %}
             <a href="{{ path('schedules_by_day', {'pid': date_list_item.servicePid, 'date': date_list_item.datetime.format('Y-m-d')}) }}"
-               class="{{ date_list_item.cssClassesForItem }}" {{ date_list_item.attributesForItem }}>
+               class="{{ item_classes }}" {{ date_list_item.attributesForItem }}>
                 {{ self.dateLines(date_list_item.isToday, date_list_item.datetime) }}
             </a>
         {% else %}
-            <span class="{{ date_list_item.cssClassesForItem }}">
+            <span class="{{ item_classes }}">
                 {{ self.dateLines(date_list_item.isToday, date_list_item.datetime) }}
             </span>
         {% endif %}
